requirements.txt
django
djangorestframework
ipython
celery
django-prometheus
